Refucktored Pressure Plate to make Base Class for all Switches
Refucktor induced (or just showed?) a bug with the interaction between
the duck and the switches:
When the Duck is picked up, the OnTriggerExit2D is not called (to me it
looks like the collision enabling/disabling should be done in a
different way).
This doesn't happend when putting down the Duck atop a switch
(OnTriggerEnter2D is called correctly).
One quick and dirty fix would be to just take into account which objects
are triggering the OnTriggerStay2D trigger in the switch, but it will
work just as long as we only need to take into account the Duck and the
Player, and will make the class not reusable later :( .